Fix for spaces pasted at start of line
When pasting in TinyMCE at the start of a paragraph, if there is a space at the start of the pasted text, the pated text is interpreted by mediawiki as a psuedo 'pre' block.  This is how mediawiki interprets text on a line preceded by a space character.
To avoid this, the space must be replaced by a non-breaking space character '&nbsp;'.